efi/api/requirements.txt
alembic==1.17.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
passlib==1.7.4
PyJWT==2.10.1
PyMySQL==1.1.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
